requirements.txt
django==5.1.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
celery==5.4.0
django-celery-beat==2.7.0
django-filter==24.3
Faker==30.0.0